requirements.txt

streamlit
google-cloud-texttospeech
pydub
librosa
numpy
scipy
